refactor(db): migrate from PostgreSQL to SQLite for MVP
- Replace postgres driver with @libsql/client (Turso-compatible)
- Rewrite all schemas: pgTable -> sqliteTable, pgEnum -> text, bigint -> integer, timestamp_ms
- Replace jsonb columns with text +  annotations
- Generate fresh SQLite migrations
- Update API: env schema, health check, audit serialization
- Update tests: temp file-based DB with auto-migration
- Remove PostgreSQL-specific SQL (::int casts, TRUNCATE)
- Remove Docker dependency for local dev (keep docker-compose.yml as reference)
- 181 tests passing, typecheck + lint clean

+++ apps/api/src/routes/auth.ts
@@ -272,7 +272,7 @@ export async function authRoutes(app: FastifyInstance) {
 
     if (existing) {
       const oneHourAgo = new Date(Date.now() - 60 * 60 * 1000)
-      const [countResult] = await db.select({ count: sql<number>`count(*)::int` })
+      const [countResult] = await db.select({ count: sql<number>`count(*)` })
         .from(auditLogs)
         .where(and(
           eq(auditLogs.userId, existing.userId),
